Updated to create correct PDF output size
When opened in Illustrator or Photoshop, PDF exports were much bigger than than the 300dpi bitmap counterparts.

Although resolution independent, their dimensions are calculated in points rather than pixels. Photoshop / Illustrator factor for this when importing PDFs.

Whilst not a major issue - I had to re-size PDFs to send to the Riso machine at 'actual size'. Correcting this removes a step of post production.

Now you input the print resolution you are working at and it will use this to calculate the scalefactor when creating PDF buffer and drawing graphics to the PDF.

Also compressed all pages that are expected to be printed with black ink (text and date readout) to a single page of the PDF. Again to reduce post production for riso prep.